[refactor] CommonSecurityConfig를 abstract 템플릿 메서드 패턴으로 변경
- 기존 concrete @Configuration → abstract 클래스로 전환
- configureRequestMatchers() 추상 메서드 추가
- 각 서비스가 상속받아 고유 URL matchers만 오버라이드
- 중복 config/CommonSecurityConfig.java 삭제
